move enum to common library
