+++ biosimulations/apps/dispatch-service/src/app/results/results.controller.ts
@@ -6,7 +6,7 @@ import {
 import { Controller, Logger } from '@nestjs/common';
 
 import { MessagePattern } from '@nestjs/microservices';
-import { SimulationRunStatus } from '@biosimulations/dispatch/api-models';
+import { SimulationRunStatus } from '@biosimulations/datamodel/common'
 import { ArchiverService } from '../services/archiver/archiver.service';
 import { ResultsService } from './results.service';
