+++ biosimulations/apps/dispatch-service/src/app/services/hpc/hpc.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, Logger, Inject } from '@nestjs/common';
 import { SshService } from '../ssh/ssh.service';
 import { ClientProxy } from '@nestjs/microservices';
-import { SimulationRunStatus } from '@biosimulations/dispatch/api-models';
+import { SimulationRunStatus } from '@biosimulations/datamodel/common'
 import { ConfigService } from '@nestjs/config';
 import { SbatchService } from '../sbatch/sbatch.service';
