+++ biosimulations/apps/dispatch/src/app/components/run/dispatch/dispatch.component.ts
@@ -10,7 +10,7 @@ import { DispatchService, SimulatorVersionsMap } from '../../../services/dispatc
 import { SimulationService } from '../../../services/simulation/simulation.service';
 import { environment } from '@biosimulations/shared/environments';
 import { Simulation } from '../../../datamodel';
-import { SimulationRunStatus } from '@biosimulations/dispatch/api-models'
+import { SimulationRunStatus } from '@biosimulations/datamodel/common'
 import { combineLatest } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { ConfigService } from '@biosimulations/shared/services';
