+++ biosimulations/apps/dispatch/src/app/datamodel.ts
@@ -1,4 +1,4 @@
-import { SimulationRunStatus } from '@biosimulations/dispatch/api-models';
+import { SimulationRunStatus } from '@biosimulations/datamodel/common';
 
 
 export interface Simulation {
