+++ biosimulations/apps/dispatch/src/app/services/simulation/simulation-status.service.ts
@@ -1,4 +1,4 @@
-import { SimulationRunStatus } from '@biosimulations/dispatch/api-models';
+import { SimulationRunStatus } from '@biosimulations/datamodel/common'
 
 
 export class SimulationStatusService {
