+++ biosimulations/apps/dispatch/src/app/services/simulation/simulation.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Simulation } from '../../datamodel';
-import { SimulationRunStatus } from '@biosimulations/dispatch/api-models'
+import { SimulationRunStatus } from '@biosimulations/datamodel/common'
 import { SimulationStatusService } from './simulation-status.service';
 import { Storage } from '@ionic/storage';
 import { HttpClient } from '@angular/common/http';
