+++ biosimulations/libs/datamodel/common/src/lib/core/index.ts
@@ -5,6 +5,7 @@ export * from './simulation-results';
 export * from './project';
 export * from './algorithm';
 export * from './simulation';
+export * from './simulationRun';
 export * from './user';
 export * from './simulator';
 export * from './dependent-package';
