+++ biosimulations/libs/datamodel/common/src/lib/core/simulation.ts
@@ -1,7 +1,7 @@
 import {
   BiomodelParameter,
   AlgorithmParameter,
-  SimulationRunAttributes,
+
   IAlgorithm,
 } from '.';
