+++ biosimulations/libs/dispatch/api-models/src/index.ts
@@ -1,3 +1,2 @@
 export * from './lib/simulationRun';
-export * from './lib/simulationRunStatus';
 export * from './lib/simulationRunReport';
